backend/requirements.txt
astapi>=0.95
uvicorn[standard]>=0.23
sqlmodel>=0.0.8
psycopg2-binary>=2.9
